requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
